window link-To convertor(test1)
after double checking that the "to_converter_button" var was correct. the partner function will allow the button to create a screen and disable it to prevent duplication and will only be alizable the button after the screen has been exited. all is needed now is to add in the detail rows of code that will be show onto the screen.