window link-convertor_frame (bolder-test5)
not exactly as I thought but master is in the collective box, A bolder frame is being produced and i think it'll fit well for this particular window. all that's needed now is to get the buttons in line with the text and input the tile frame with the inputs text box.